+++ awb_aradial_connector/api/aradial_gateway/get_user.py
@@ -22,10 +22,8 @@ class AradialAPIGatewayGetUser(object):
     def get_user(self, sms_id_username):
 
         try:
-            url = self.url + "/" + sms_id_username
-            _logger.info(f'API URL: {url}')
             res = requests.get(
-                url=url,
+                url=self.url + "/" + sms_id_username,
                 auth=HTTPBasicAuth(self.username, self.password)
             )
         except requests.exceptions.MissingSchema as e: