+++ awb_generic_sms/docs/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.3.6] - 2021-09-20
+   - [FIX] Updated SMS template as per the Spiels Matrix
+
+## [13.0.1.3.5] - 2021-09-16
+   - [MISC] Added debug logs for troubleshooting
+
 ## [13.0.1.3.4] - 2021-09-14
    - Welcome Notification enhancements
    - Activation Notification enhancements