+++ awb_subscriber_bill_automation/CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.5.5] - 2021-09-20
+- [FIX] Update opportunity id of new subscription
+
+## [13.0.1.5.4] - 2021-09-20
+- [FIX] Populate PrepaidIndicator flag in the Aradial Create User payload
+- [MISC] Proper logging for the debug logs
+- [FIX] SF API Update Account trigger
+- [ADD] Update of latest subscription upon reloading
+
+
+## [13.0.1.5.3] - 2021-09-16
+- [FIX] Handling for Main Plan in Postpaid subscription
+- [FIX] Error in SMS Sending
+- [ADD] Debug Logs
+
+## [13.0.1.5.2] - 2021-09-15
+- [ADD] SF API integration and trigger in creation of Subscription
+
+## [13.0.1.5.1] - 2021-09-15
+- [FIX] Added missing fields (Removed last night)
+- [ADD] New field datetime for the TMS Notification
+
 ## [13.0.1.5.0] - 2021-09-08
 - [IMP] SMS Provisioning and Activation flows
 - [ADD] Disconnection API